Fix two numeric representation bugs
1) In CodePrinter: Printing of some large numbers
above 2^53 were incorrectly truncated.
2) In NewIRFactory: Literal hex, binary, and octal
numeric literal representations couldn't be used for literals
larger than the maximum representable long (2^63-1).

Fixes #691 on github
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=79345667

+++ src/com/google/javascript/jscomp/CodeConsumer.java
@@ -256,7 +256,7 @@ abstract class CodeConsumer {
       long mantissa = value;
       int exp = 0;
       if (Math.abs(x) >= 100) {
-        while (mantissa / 10 * Math.pow(10, exp + 1) == value) {
+        while (mantissa / 10 * ((long) Math.pow(10, exp + 1)) == value) {
           mantissa /= 10;
           exp++;
         }
